\name{myCluster-package}
\alias{myCluster-package}
\alias{myCluster}
\docType{package}
\title{
\packageTitle{myCluster}
}
\description{
\packageDescription{myCluster}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{myCluster}
\packageIndices{myCluster}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{myCluster}

Maintainer: \packageMaintainer{myCluster}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
